+++ src/android/com/appsflyer/cordova/plugin/AppsFlyerConstants.java
@@ -42,4 +42,15 @@ public class AppsFlyerConstants {
     final static String INVITE_CUSTOMERID = "customerID";
     final static String INVITE_DEEPLINK = "baseDeepLink";
     final static String PROMOTE_ID = "promotedAppId";
+
+    //RECEIPT VALIDATION
+    final static String PUBLIC_KEY = "publicKey";
+    final static String SIGNATURE = "signature";
+    final static String PURCHASE_DATA = "purchaseData";
+    final static String PRICE = "price";
+    final static String CURRENCY = "currency";
+    final static String ADDITIONAL_PARAMETERS = "additionalParameters";
+    final static String NO_PARAMETERS_ERROR = "Please provide purchase parameters";
+    final static String VALIDATE_SUCCESS = "In-App Purchase Validation success";
+    final static String VALIDATE_FAILED = "In-App Purchase Validation failed with error: ";
 }